+++ craft_parts/packages/base.py
@@ -1,6 +1,6 @@
 # -*- Mode:Python; indent-tabs-mode:nil; tab-width:4 -*-
 #
-# Copyright 2017-2023 Canonical Ltd.
+# Copyright 2017-2025 Canonical Ltd.
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU Lesser General Public
@@ -212,6 +212,7 @@ class DummyRepository(BaseRepository):
         refresh_package_cache: bool = True,  # noqa: ARG003
     ) -> list[str]:
         """Install packages on the host system."""
+        logger.debug("Package manager not defined, not installing any packages")
         return []
 
     @classmethod